refactor : todo/text 책임 분리 (트랜잭션)
