+++ src/main/java/org/onesentence/onesentence/domain/user/service/UserService.java
@@ -12,4 +12,6 @@ public interface UserService {
 
 	User findByUserId(Long userId);
 
+	void checkUserExistence(Long userId);
+
 }
